Move InputValInfo & EquatableGType to Common
